backend/requirements.txt
fastapi
uvicorn
pydantic
aiosqlite
loguru
python-dotenv
requests
